Heavy improvement of performance and fixing issue #115
the canConnect and canContain callbacks are now not called anymore.

the static stencil set rules are now cached at load time. this dramatically increases the load time by ~15seconds on Mac, but on the other hand heavily increases the performance on run time.

Dramatically decreased the number of calls of the update mechanism. The number of calls decreased by ~80%.



git-svn-id: http://oryx-editor.googlecode.com/svn/trunk@474 d672c736-503d-0410-a38a-9366997c882b