fixed a bug that came up when importing an erdf resource. There was the problem of conflicting ids and a broken conflict resolution. Now, the editor uses UUIDs to guarantee uniqueness and to avoid conflicts.
the importJSON method now renews all ids of the imported objects to avoid any conflicts.

also fixed a problem of mispositioned bpmn intermediate events after exporting and then importing again. the problem was the position of the docker to attach the event was also serialized when the event was not attached.

git-svn-id: http://oryx-editor.googlecode.com/svn/trunk@990 d672c736-503d-0410-a38a-9366997c882b